requirements-docker.txt
boto3==1.10.41
botocore==1.13.41
Click==7.0
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
Flask==1.1.1
gunicorn==20.0.4
imageio==2.6.1
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
pyparsing==2.4.5
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
s3transfer==0.2.1
scipy==1.3.3
six==1.13.0
torchvision
urllib3==1.24.3
Werkzeug==0.16.0
